Split SDL main thread class into its own file
